sdk/pyproject.toml
[tool.poetry]
name = "honcho-ai"
version = "0.0.0.dev1"
description = "Python Client SDK for Honcho"
authors = ["Plastic Labs <hello@plasticlabs.ai>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "honcho"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
tqdm = "^4.66.1"
rich = "^13.7.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
